+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2023-02-01
+
+- Makes the `values` field protected, use `getValues()` instead.
+- Makes the `Row` class iterable.
+- Adds `Row` metadata functions.
+
 ## [0.1.1] - 2023-01-30
 
 - Fixes broken composer.lock file.